# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Subcults API
  description: |
    Privacy-first platform for mapping underground music communities.
    Combines real-time location awareness, live audio streaming, and trust-based discovery.

    ## Authentication
    JWT-based authentication with two token types:
    - **Access tokens**: 15-minute expiry, includes `did` (Decentralized Identifier) claim
    - **Refresh tokens**: 7-day expiry, cookie-only

    Pass the access token via `Authorization: Bearer <token>` header or the `access_token` cookie.

    ## Privacy
    All public location data uses geohash-based jitter to prevent precise location tracking.
    Precise coordinates are only available when `allow_precise` consent is granted.

    ## Rate Limiting
    Per-endpoint rate limits are enforced. When exceeded, the API returns `429 Too Many Requests`
    with a `Retry-After` header.
  version: 0.1.0
  contact:
    name: Subcults
  license:
    name: MIT
servers:
  - url: /
    description: Current deployment
...
paths:
...
  /search/appearances:
    get:
      operationId: searchAppearances
      tags: [Touring, Search]
      summary: Search public appearances
      security: [{}, {bearerAuth: []}]
      parameters:
        - {name: place, in: query, schema: {type: string}}
        - {name: bbox, in: query, schema: {type: string}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
        - {name: act, in: query, schema: {type: string}}
        - {name: tour, in: query, schema: {type: string}}
        - {name: festival, in: query, schema: {type: boolean}}
        - {name: scene, in: query, schema: {type: string}}
        - {name: kind, in: query, schema: {type: string, enum: [show, festival, party, meetup, broadcast, other]}}
        - {name: locality, in: query, schema: {type: string, enum: [any, local, visiting]}}
        - {name: access, in: query, schema: {type: string, enum: [public, all]}}
      responses:
        '200':
          description: Public appearance results
          content: {application/json: {schema: {$ref: '#/components/schemas/TouringDetailResponse'}}}
        '400': {$ref: '#/components/responses/ValidationError'}
...
  /tours/{id}:
    get:
      operationId: getTour
      tags: [Touring]
      summary: Get a tour itinerary
      security: [{}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ResourceID'}]
      responses:
        '200': {description: Tour itinerary, content: {application/json: {schema: {$ref: '#/components/schemas/TouringDetailResponse'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT access token. Include as Authorization: Bearer <token> header or access_token cookie.'
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - validation_error
                - auth_failed
                - not_found
                - rate_limited
                - internal_error
                - forbidden
                - conflict
                - bad_request
                - invalid_palette
                - scene_deleted
                - scene_not_found
                - invalid_scene_name
                - duplicate_scene_name
                - invalid_time_range
                - missing_target
                - unsupported_type
                - invalid_weight
                - alliance_deleted
                - self_alliance
                - already_onboarded
                - not_onboarded
            message:
              type: string
              description: Human-readable error message

    # ── Geographic ──────────────────────────────────────────────────
    Point:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
...
    PublicOccurrence:
      type: object
      required: [coarse_geohash, precision]
      properties:
        coarse_geohash: {type: string}
        display_point: {$ref: '#/components/schemas/Point'}
        precision: {type: string, enum: [coarse, precise]}
    AppearanceSummary:
      type: object
      required: [id, event, act, host_names, context, locality, status, verification]
      properties:
        id: {type: string}
        event: {type: object, required: [id, title, starts_at, kind, occurrence], properties: {id: {type: string}, title: {type: string}, starts_at: {type: string, format: date-time}, kind: {type: string}, occurrence: {$ref: '#/components/schemas/PublicOccurrence'}}}
        act: {type: object, required: [id, name], properties: {id: {type: string}, name: {type: string}, home_territory: {type: string}}}
        tour: {type: object, properties: {id: {type: string}, title: {type: string}}}
        host_names: {type: array, items: {type: string}}
        context: {type: string, enum: [tour_stop, festival_appearance, one_off]}
        locality: {type: string, enum: [local, visiting, unknown]}
        status: {type: string, enum: [announced, confirmed, cancelled, completed]}
        verification: {type: string, enum: [unverified, claimed, verified, disputed, rejected]}
    TouringDetailResponse:
      type: object
      required: [appearances]
      properties:
        profile: {type: object, properties: {id: {type: string}, name: {type: string}, home_territory: {type: string}}}
        tour: {type: object, properties: {id: {type: string}, title: {type: string}}}
        appearances: {type: array, items: {$ref: '#/components/schemas/AppearanceSummary'}}

    # ── Payment ─────────────────────────────────────────────────────
...